Unmechanical — Update authors list + Update script to 2.10 syntax

+++ play.it-2/games/play-unmechanical.sh
@@ -3,6 +3,7 @@ set -o errexit
 
 ###
 # Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2018, Solène Huault
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,18 +31,16 @@ set -o errexit
 
 ###
 # Unmechanical
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20181005.1
 
 # Set game-specific variables
 
 GAME_ID='unmechanical'
 GAME_NAME='Unmechanical'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='gog_unmechanical_2.1.0.4.sh'
 ARCHIVE_GOG_URL='https://www.gog.com/game/unmechanical'
@@ -51,21 +50,19 @@ ARCHIVE_GOG_VERSION='2.0-gog2.1.0.4'
 ARCHIVE_GOG_TYPE='mojosetup'
 
 ARCHIVE_DOC_DATA_PATH='data/noarch/docs'
-ARCHIVE_DOC_DATA_FILES='./*'
+ARCHIVE_DOC_DATA_FILES='*'
 
 ARCHIVE_GAME_BIN_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN_FILES='./Binaries ./Engine'
+ARCHIVE_GAME_BIN_FILES='Binaries Engine'
 
 ARCHIVE_GAME_DATA_PATH='data/noarch/game'
-ARCHIVE_GAME_DATA_FILES='./UDKGame ./UnmechanicalIcon.bmp'
+ARCHIVE_GAME_DATA_FILES='UDKGame UnmechanicalIcon.bmp'
 
 APP_MAIN_TYPE='native'
 APP_MAIN_PRERUN='pulseaudio --start'
 APP_MAIN_EXE='Binaries/Linux/UDKGame-Linux'
 APP_MAIN_LIBS='lib'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='UnmechanicalIcon.bmp'
-APP_MAIN_ICON_RES='256'
 
 PACKAGES_LIST='PKG_BIN PKG_DATA'
 
@@ -77,40 +74,41 @@ PKG_BIN_DEPS="$PKG_DATA_ID glibc libstdc++ glu xcursor openal sdl2 pulseaudio"
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.10'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: ${XDG_DATA_HOME:="$HOME/.local/share"}
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
+fi
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
 fi
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"  "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
+# Get icon
 
 PKG='PKG_DATA'
-res="$APP_MAIN_ICON_RES"
-PATH_ICON="$PATH_ICON_BASE/${res}x${res}/apps"
-extract_icon_from "${PKG_DATA_PATH}${PATH_GAME}/$APP_MAIN_ICON"
-mkdir --parents "${PKG_DATA_PATH}${PATH_ICON}"
-mv "$PLAYIT_WORKDIR/icons/$(basename ${APP_MAIN_ICON%.bmp}.png)" "${PKG_DATA_PATH}${PATH_ICON}/$GAME_ID.png"
-
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
@@ -119,15 +117,14 @@ write_launcher 'APP_MAIN'
 
 # Set working directory to the directory containing the game binary before running it
 
-pattern1='s|^cd "$PATH_PREFIX"$|cd "$PATH_PREFIX/${APP_EXE%/*}"|'
-pattern2='s|^"\./$APP_EXE"|"./${APP_EXE##*/}"|'
+pattern='s|^cd "$PATH_PREFIX"$|cd "$PATH_PREFIX/${APP_EXE%/*}"|'
+pattern="$pattern"';s|^"\./$APP_EXE"|"./${APP_EXE##*/}"|'
 file="${PKG_BIN_PATH}${PATH_BIN}/$GAME_ID"
-sed --in-place "$pattern1;$pattern2" "$file"
+sed --in-place "$pattern" "$file"
 
 # Build package
 
-write_metadata 'PKG_DATA'
-write_metadata 'PKG_BIN'
+write_metadata
 build_pkg
 
 # Clean up
